unsignedBigInteger user_id

+++ database/migrations/2023_09_21_133017_create_education_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('education_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->string('degree')->nullable();
             $table->integer('passing_year')->nullable();
             $table->integer('country_id')->nullable();
